Updated PDF, added Hayame, and added issue tracking to technologies
Both English and Dutch are up-to-date now.

Also added an issue tracking section to technologies, this contains the following:

Jira, GitHub Issues, Space, Trello

I have also added Hayame as project, it mentions the following in the description:
SQL, Prometheus, MariaDB, HikariCP and localisation.

Made a minor adjustment to the description of my internship at Vanderlande

```
Working on an internal tool surrounding CI/CD, this tool is written in Java.
```
Added `this tool is written in Java` and rephrased to `Working`